media: dvb: mpq: expose TSPP descriptor size as module parameter
Default TSPP descriptor size is 192 bytes (1 TS packet). For high
bitrate streams this seems to cause the tasklet loop to collect a lot
of descriptors in each iteration (over 170), causing a long period
of time with a spinlock locked (even up to ~20ms) resulting in overflow.
Exposing the descriptor size as a parameter will allow to tune it
according to need and help avoid overflow situations.

Change-Id: I03d7c51024f82f506375974efb2cd0baad03111d
Signed-off-by: Gilad Broner <gbroner@codeaurora.org>