ARM: dts: msm: Fix for GPIO conflict issue on MSM8996 and APQ8096
Create individual node for MI2S tert SD1 so that it will only be
included in the APQ8096 MDM9x55 I2S targets and not conflict with
MSM8996. Also, add fix for pinctrl GPIO issue that was causing
the codec read and writes to fail. Add support for QUAT SD1 GPIO.

CRs-fixed: 1024729
Change-Id: Ib687057c1a72d0b707eab711d17571bf8b1ff474
Signed-off-by: Josh Kirsch <jkirsch@codeaurora.org>