msm: vidc: Initialize the dcvs load during perf mode switch
Incase V4L2 client sets the perf mode to TURBO, DCVS
is turned off. Now when the client sets the perf mode
back to nominal, venus clock and load is changed but
dcvs parameters are not updated to reflect the driver
current state interms of venus load. Initialize the
dcvs parameters to update itself as per the ongoing
venus load.

Change-Id: I8f35cdd154fbde49c95cdd0e5f7df4b9e4f897a5
Signed-off-by: Vikash Garodia <vgarodia@codeaurora.org>